feat: Refactor JS for new module
